Added test code to RangeServer to update METADATA range;
This code (all under #ifdef METADATA_TEST) will set a global variable
to contain a RangeServerClient for the second-level METADATA
range if it is loaded.  Subsequently all METADATA updates will get
written to this Range in addition to the mock metadata.  Also fixed
a bug in ConnectionManager where it was deleting itself when
it lost its one and only connection


git-svn-id: https://hypertable.googlecode.com/svn/trunk@423 01261837-d133-0410-8dcb-6792eb736156