Fixed race in RangeServer::update added ROW_KEY_COLUMN= to LOAD DATA INFILE
There was a race condition inside RangeServer::update that was causing the
regression to intermittently fail.  Added ROW_KEY_COLUMN= option to
the LOAD DATA INFILE command.  The column can be specified as an identifier
or a string literal (e.g. with quotes if the name of the column conflicts with
an HQL keyword).


git-svn-id: https://hypertable.googlecode.com/svn/trunk@501 01261837-d133-0410-8dcb-6792eb736156