Changed CellCache so that they can share entries.
Changed the map so that the value is a 32-bit offset from the
beginning of the key and bit 31 is a flag that indicates whether
or not to deallocate the k/v pair upon destruction.  During
a slice_copy, the parent indicates to the child (the new cell cache)
that it is referencing it.  The child waits in the destructor
until the parent dies before deallocating it's k/v pairs.