Cleaned up start/stop scripts
Changed start-test-servers.sh to start-all-servers.sh and removed the
creation of the test tables.  Creating the test tables is now a
separate step required before running the regressions.  Modified
dfsclient to exit immediately with an error message if unable to
connect to the DfsBroker.  Renamed stop-servers.sh to kill-servers.sh