Changed CellStore block size to be 65K of compressed data.
Before it was 65K of uncompressed data which was causing small
blocks to be created.  Also changed the exponentially incresing
max range size to start out smaller initially.  Added a retry
in the MutatorScatterBuffer for PARTIAL updates.  Also got
rid of local commit log for Hadoop operation in the startup
scripts.