Fixed METADATA table creation/reference problem
The creation of the METADATA table is now done lazily
after the first RangeServer is registered.  The RangeServers
now lazily open the METADATA table upon the first call
to load_range.  Fixed an uninitialized data problem in
the final Header in the commit log.  Also changed the default
compressor for the Commit log from 'quicklz' to 'lzo'