Moved header checksum to end of header; Improved split row selection
Moved the header checksum of the BlockCompressionHeader classes
to the end of the header and changed it to be the top 2 bytes
of the fletcher32.  Also made a small improvement in the
split row determination logic inside Range;  Now if the range
only consists of two rows, the first row will be selected
as the split row.